requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
bottle>=0.12.25
